COZMO-15263: Code Lab: Magic Fortune errors on open and delete (#6228)
Bugs encountered and fixed include:

1. scratch-blocks throws error from xml.js method Blockly.Xml.domToWorkspace when loading the project: "Shadow block cannot be a top-level block." The block causing this issue is a sensor (like a Cozmo sensor or even the "current date" sensor) inside a greater than operator. Both scratch-blocks and scratch-vm playgrounds don't have this error. I proved that for this serialized project JSON, scratch-blocks does not go through the else statement I commented out in xml.js in this commit.

2. scratch-vm throws an error on deleting the whole stack of blocks in blocks.js method deleteBlock. Now we record the error and return from deleteBlock, and this allows the rest of the stack to be deleted.

With these changes, COZMO-15263 is fixed and the user can successfully delete the stack of blocks and create a remixed project with whatever new blocks they want.

+++ unity/Cozmo/Assets/StreamingAssets/Scratch/lib/vm/src/engine/blocks.js
@@ -408,6 +408,14 @@ class Blocks {
         // Get block
         const block = this._blocks[e.id];
 
+        // *** ANKI CHANGE ***
+        // We shouldn't have an undefined block, but if we do, return
+        // so that the work to delete can complete.
+        if (block == undefined) {
+          window.cozmoDASLog("Codelab.ScratchVM.DeleteBlock.BlockUndefined", "In blocks.js deleteBlock, block is undefined so returning early.");
+          return;
+        } 
+
         // Delete children
         if (block.next !== null) {
             this.deleteBlock({id: block.next});
